requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
scikit-learn>=1.3
statsmodels>=0.14
fredapi>=0.5
python-dotenv>=1.0
requests>=2.31
typer[all]>=0.9
rich>=13.6
streamlit>=1.28
plotly>=5.15
prophet>=1.1
tensorflow>=2.13
torch>=2.0
seaborn>=0.12
yfinance>=0.2
tweepy>=4.14
newspaper3k>=0.2.8
openai>=1.0
transformers>=4.35
pymc>=5.10
arviz>=0.16
ruptures>=1.1
networkx>=3.2
feedparser>=6.0
scipy>=1.11